+++ Exodus/src/utopia/exodus/database/factory/user/SessionFactory.scala
@@ -6,6 +6,7 @@ import utopia.exodus.model.partial.UserSessionData
 import utopia.exodus.model.stored.UserSession
 import utopia.flow.datastructure.immutable.{Constant, Model}
 import utopia.flow.time.Now
+import utopia.metropolis.model.enumeration.ModelStyle
 import utopia.vault.model.enumeration.ComparisonOperator.Larger
 import utopia.vault.nosql.factory.{Deprecatable, FromValidatedRowModelFactory}
 
@@ -24,7 +25,7 @@ object SessionFactory extends FromValidatedRowModelFactory[UserSession] with Dep
 	
 	override protected def fromValidatedModel(model: Model[Constant]) = UserSession(model("id").getInt,
 		UserSessionData(model("userId").getInt, model("key").getString, model("expiresIn").getInstant,
-			model("deviceId").int))
+			model("deviceId").int, model("modelStylePreference").int.flatMap(ModelStyle.forId)))
 	
 	override def table = ExodusTables.userSession