+++ Metropolis/src/utopia/metropolis/model/combined/language/DescribedLanguage.scala
@@ -1,6 +1,6 @@
 package utopia.metropolis.model.combined.language
 
-import utopia.metropolis.model.combined.description.{Described, DescribedFromModelFactory}
+import utopia.metropolis.model.combined.description.{DescribedWrapper, DescribedFromModelFactory}
 import utopia.metropolis.model.stored.description.DescriptionLink
 import utopia.metropolis.model.stored.language.Language
 
@@ -14,7 +14,7 @@ object DescribedLanguage extends DescribedFromModelFactory[DescribedLanguage, La
   * @author Mikko Hilpinen
   * @since 16.5.2020, v1
   */
-case class DescribedLanguage(language: Language, descriptions: Set[DescriptionLink]) extends Described[Language]
+case class DescribedLanguage(language: Language, descriptions: Set[DescriptionLink]) extends DescribedWrapper[Language]
 {
 	override def wrapped = language
 }