+++ Metropolis/src/utopia/metropolis/model/combined/language/DescribedLanguageFamiliarity.scala
@@ -1,6 +1,6 @@
 package utopia.metropolis.model.combined.language
 
-import utopia.metropolis.model.combined.description.{Described, DescribedFromModelFactory}
+import utopia.metropolis.model.combined.description.{DescribedWrapper, DescribedFromModelFactory}
 import utopia.metropolis.model.stored.description.DescriptionLink
 import utopia.metropolis.model.stored.language.LanguageFamiliarity
 
@@ -15,4 +15,4 @@ object DescribedLanguageFamiliarity extends DescribedFromModelFactory[DescribedL
   * @since 25.7.2020, v1
   */
 case class DescribedLanguageFamiliarity(wrapped: LanguageFamiliarity, descriptions: Set[DescriptionLink])
-	extends Described[LanguageFamiliarity]
+	extends DescribedWrapper[LanguageFamiliarity]