+++ Metropolis/Changes.md
@@ -1,6 +1,12 @@
 # Utopia Metropolis - List of Changes
 
-## v1.0.1 (in development)
+## v1.1 (in development)
+### Breaking Changes
+- Renamed the **Described** trait to **DescribedWrapper**
+### New Features
+- Added simple model support (**ModelStyle** and **StyledModelConvertible**)
+    - Also added simple styling support for described items 
+      (using **SimplyDescribed** and **SimplyDescribedWrapper** traits)
 ### Other Changes
 - Added a few models from the **Exodus Module**
 - **DeletionCancel** now contains property `.created`